fix MPS-22180 Wrong scopes in completion menu for 'execute after', 'required data', 'produced data', 'description' in migration class: minor fix in editor

+++ languages/languageDesign/classLike/source_gen/jetbrains/mps/lang/classLike/editor/EditorUtil.java
@@ -17,16 +17,19 @@ public class EditorUtil {
     SNodeOperations.replaceWithAnother(ph, BehaviorReflection.invokeVirtual((Class<SNode>) ((Class) Object.class), SLinkOperations.getTarget(ph, MetaAdapterFactory.getReferenceLink(0xc7d5b9dda05f4be2L, 0xbc73f2e16994cc67L, 0x59e9926e840d7db2L, 0x59e9926e840d7db5L, "decl")), "virtual_create_8260330507834998478", new Object[]{}));
   }
   public static void restorePlaceholder(final SNode mi) {
-    if (!((ClassLikeMember_Behavior.call_isRequired_3402736933911994098(BehaviorReflection.invokeVirtual((Class<SNode>) ((Class) Object.class), mi, "virtual_getDeclaration_9097849371503884215", new Object[]{})) && Sequence.fromIterable(SNodeOperations.ofConcept(SLinkOperations.getChildren(SNodeOperations.getNodeAncestor(mi, MetaAdapterFactory.getConcept(0xf3061a5392264cc5L, 0xa443f952ceaf5816L, 0xf8c108ca66L, "jetbrains.mps.baseLanguage.structure.ClassConcept"), false, false), MetaAdapterFactory.getContainmentLink(0xf3061a5392264cc5L, 0xa443f952ceaf5816L, 0x101d9d3ca30L, 0x4a9a46de59132803L, "member")), MetaAdapterFactory.getInterfaceConcept(0xc7d5b9dda05f4be2L, 0xbc73f2e16994cc67L, 0x59e9926e840d9179L, "jetbrains.mps.lang.classLike.structure.ClassLikeMemberInstance"))).where(new IWhereFilter<SNode>() {
+    boolean lastInstance = ClassLikeMember_Behavior.call_isRequired_3402736933911994098(BehaviorReflection.invokeVirtual((Class<SNode>) ((Class) Object.class), mi, "virtual_getDeclaration_9097849371503884215", new Object[]{})) && Sequence.fromIterable(SNodeOperations.ofConcept(SLinkOperations.getChildren(SNodeOperations.getNodeAncestor(mi, MetaAdapterFactory.getConcept(0xf3061a5392264cc5L, 0xa443f952ceaf5816L, 0xf8c108ca66L, "jetbrains.mps.baseLanguage.structure.ClassConcept"), false, false), MetaAdapterFactory.getContainmentLink(0xf3061a5392264cc5L, 0xa443f952ceaf5816L, 0x101d9d3ca30L, 0x4a9a46de59132803L, "member")), MetaAdapterFactory.getInterfaceConcept(0xc7d5b9dda05f4be2L, 0xbc73f2e16994cc67L, 0x59e9926e840d9179L, "jetbrains.mps.lang.classLike.structure.ClassLikeMemberInstance"))).where(new IWhereFilter<SNode>() {
       public boolean accept(SNode it) {
         return BehaviorReflection.invokeVirtual((Class<SNode>) ((Class) Object.class), it, "virtual_getDeclaration_9097849371503884215", new Object[]{}) == BehaviorReflection.invokeVirtual((Class<SNode>) ((Class) Object.class), mi, "virtual_getDeclaration_9097849371503884215", new Object[]{});
       }
-    }).count() == 1))) {
-      if ((ClassLikeMember_Behavior.call_getPlaceholder_9097849371503188814(BehaviorReflection.invokeVirtual((Class<SNode>) ((Class) Object.class), mi, "virtual_getDeclaration_9097849371503884215", new Object[]{})) != null)) {
-        SNodeOperations.replaceWithAnother(mi, PlaceholderModifier_Behavior.call_create_9097849371503335421(ClassLikeMember_Behavior.call_getPlaceholder_9097849371503188814(BehaviorReflection.invokeVirtual((Class<SNode>) ((Class) Object.class), mi, "virtual_getDeclaration_9097849371503884215", new Object[]{}))));
-      } else {
-        SNodeOperations.deleteNode(mi);
-      }
+    }).count() == 1;
+    if (lastInstance) {
+      return;
+    }
+
+    if ((ClassLikeMember_Behavior.call_getPlaceholder_9097849371503188814(BehaviorReflection.invokeVirtual((Class<SNode>) ((Class) Object.class), mi, "virtual_getDeclaration_9097849371503884215", new Object[]{})) != null)) {
+      SNodeOperations.replaceWithAnother(mi, PlaceholderModifier_Behavior.call_create_9097849371503335421(ClassLikeMember_Behavior.call_getPlaceholder_9097849371503188814(BehaviorReflection.invokeVirtual((Class<SNode>) ((Class) Object.class), mi, "virtual_getDeclaration_9097849371503884215", new Object[]{}))));
+    } else {
+      SNodeOperations.deleteNode(mi);
     }
   }
 }
